Refactor lambda handler to use scales package
